sistema_de_controle_de_estoque/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-mysql-connector==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.1.2
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.0.4
